+++ app/src/main/java/com/codesoom/assignment/controllers/ProductErrorAdvice.java
@@ -1,0 +1,27 @@
+package com.codesoom.assignment.controllers;
+
+import com.codesoom.assignment.dto.ErrorResponse;
+import com.codesoom.assignment.errors.ProductNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+/**
+ * 상품 관련 예외처리를 담당합니다.
+ */
+@ResponseBody
+@ControllerAdvice
+public class ProductErrorAdvice {
+
+    /**
+     * 상품을 찾지 못 할 때 에러 응답을 리턴합니다.
+     * @return 에러응답
+     */
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(ProductNotFoundException.class)
+    public ErrorResponse handleProductNotFound() {
+        return new ErrorResponse("Product not found");
+    }
+}
