+++ app/src/main/java/com/codesoom/assignment/controllers/UserErrorAdvice.java
@@ -1,7 +1,8 @@
 package com.codesoom.assignment.controllers;
 
 import com.codesoom.assignment.dto.ErrorResponse;
-import com.codesoom.assignment.errors.*;
+import com.codesoom.assignment.errors.UserEmailDuplicationException;
+import com.codesoom.assignment.errors.UserNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -12,39 +13,38 @@ import javax.validation.ConstraintViolation;
 import javax.validation.ConstraintViolationException;
 import java.util.Set;
 
+/**
+ * 사용자 관련 예외처리를 담당합니다.
+ */
 @ResponseBody
 @ControllerAdvice
-public class ControllerErrorAdvice {
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(ProductNotFoundException.class)
-    public ErrorResponse handleProductNotFound() {
-        return new ErrorResponse("Product not found");
-    }
+public class UserErrorAdvice {
 
+    /**
+     * 사용자를 찾지 못 했을때 에러 응답을 리턴합니다.
+     * @return 에러 응답
+     */
     @ResponseStatus(HttpStatus.NOT_FOUND)
     @ExceptionHandler(UserNotFoundException.class)
     public ErrorResponse handleUserNotFound() {
         return new ErrorResponse("User not found");
     }
 
+    /**
+     * 사용자 이메일이 중복 될 때 에러 응답을 리턴합니다.
+     * @return 에러 응답
+     */
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(UserEmailDuplicationException.class)
     public ErrorResponse handleUserEmailIsAlreadyExisted() {
         return new ErrorResponse("User's email address is already existed");
     }
 
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(LoginFailException.class)
-    public ErrorResponse handleLoginFailException() {
-        return new ErrorResponse("Log-in failed");
-    }
-
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    @ExceptionHandler(InvalidTokenException.class)
-    public ErrorResponse handleInvalidAccessTokenException() {
-        return new ErrorResponse("Invalid access token");
-    }
-
+    /**
+     * DTO로 바인딩 실패 할 때 에러 응답을 리턴합니다.
+     * @param exception
+     * @return
+     */
     @ResponseBody
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(ConstraintViolationException.class)
